+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ## [0.0.3.5]
 2025-05
 
+> [!NOTE]
 > Namechange from Image-Portfolio to LensFlow
 
 ### New Features