+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 > [!NOTE]
-> You finde a complete Changelog here .changelog/Changelog.md
+> You find a complete Changelog here .changelog/Changelog.md
 
 ## [0.0.3.5]
 2025-05