+++ CHANGELOG.md
@@ -6,10 +6,13 @@
 ## [0.0.3.5]
 2025-05
 
+**Full Changelog**: https://github.com/cheinisch/Framora/compare/v0.0.3.4...v0.0.3.5
+
 > [!NOTE]
 > Namechange from Image-Portfolio to Framora
 
 ### New Features
+- Add Logintype "One TIme Password via Mail"
 
 ## [0.0.3.4]
 2025-04-30